Refactor graph internals to remove facade-driven dependencies, centralize auditor/fan-out invariants, retarget tests to owning modules, and refresh orientation docs around the current package architecture
