--- docs/INDEX.md
@@ -1,156 +1,0 @@
-# Documentation Index
-
-Welcome to the Investment Agent documentation! This directory contains guides, tutorials, and reference materials for using and extending the multi-agent investment analysis system.
-
-## 📚 Documentation Structure
-
-Documentation is organized by audience and purpose:
-
-### For New Users
-
-- **[Getting Started](getting-started/)** - First-time setup and basic usage *(coming soon)*
-  - Installation guide
-  - Your first analysis
-  - Configuration options
-
-- **[User Guide](user-guide/)** - Day-to-day usage *(coming soon)*
-  - Batch analysis
-  - Interpreting results
-  - Troubleshooting
-
-### For Developers
-
-- **[Developer Guide](developer-guide/)** - Extending the system *(coming soon)*
-  - Architecture overview
-  - Adding new agents
-  - Integrating data sources
-  - Testing practices
-
-- **[Examples](examples/)** - Practical code examples *(coming soon)*
-  - Custom investment thesis
-  - Sector-specific analysis
-  - Adding new tools
-
-### For Learning Agentic AI
-
-- **[Agentic AI](agentic-ai/)** - Educational content *(coming soon)*
-  - Introduction to agentic AI
-  - Multi-agent patterns
-  - Code walkthrough
-  - Framework comparisons
-
-### Infrastructure
-
-- **[Deployment](deployment/)** - Production setup *(coming soon)*
-  - Docker guide
-  - Terraform/cloud deployment
-  - CI/CD configuration
-
-- **[Security](security/)** - Security practices *(coming soon)*
-  - Vulnerability scanning
-  - Secret management
-  - Best practices
-
-## 🚀 Quick Links
-
-- **[README.md](../README.md)** - Project overview and quick start
-- **[CODEBASE_MEMORY.md](./CODEBASE_MEMORY.md)** - Durable codebase orientation snapshot
-- **[CLAUDE.md](../CLAUDE.md)** - Comprehensive developer guide
-- **[CONTRIBUTING.md](../CONTRIBUTING.md)** - How to contribute
-- **[CHANGELOG.md](../CHANGELOG.md)** - Version history
-
-## 📖 Currently Available
-
-While the full documentation structure is being built, these resources are currently available:
-
-### Core Documentation (Root Directory)
-
-1. **[README.md](../README.md)** - Complete project overview
-   - Architecture diagrams
-   - Installation and setup
-   - Quick start guide
-   - Investment thesis explanation
-   - Batch analysis instructions
-   - Performance benchmarks
-
-2. **[CODEBASE_MEMORY.md](./CODEBASE_MEMORY.md)** - Maintained orientation brief
-   - Current branch and release context
-   - Runtime architecture and subsystem map
-   - History summary from changelog and git
-   - Recommended reorientation path for future sessions
-
-3. **[CLAUDE.md](../CLAUDE.md)** - Developer reference
-   - Essential commands
-   - Architecture deep-dive
-   - File-by-file breakdown
-   - Common development patterns
-   - Testing strategy
-   - Troubleshooting guide
-
-4. **[CONTRIBUTING.md](../CONTRIBUTING.md)** - Contribution guidelines
-   - Development setup
-   - Code quality standards
-   - Testing requirements
-   - Pull request process
-
-### Examples
-
-- **[examples/analyze_single_ticker.sh](../examples/analyze_single_ticker.sh)** - Minimal example with setup validation
-- **[examples/check_token_costs.py](../examples/check_token_costs.py)** - Token usage and cost estimation
-
-### Scripts
-
-- **[scripts/run_tickers.sh](../scripts/run_tickers.sh)** - Batch analysis script
-- **[scripts/SCRIPTS_QUICK_REFERENCE.md](../scripts/SCRIPTS_QUICK_REFERENCE.md)** - Script documentation
-
-## 🔨 Under Construction
-
-The following documentation sections are planned and will be added progressively:
-
-### High Priority (Coming Soon)
-
-- [ ] **Getting Started Tutorial** - Step-by-step first analysis
-- [ ] **Agentic AI 101** - Introduction to multi-agent systems
-- [ ] **Code Walkthrough** - Annotated flow of an analysis
-- [ ] **Build Your Own Agent** - Tutorial for adding agents
-
-### Medium Priority
-
-- [ ] **Architecture Guide** - System design and patterns
-- [ ] **Interpreting Results** - How to read analysis reports
-- [ ] **Investment Thesis** - GARP strategy deep-dive
-- [ ] **Data Sources** - API integration guide
-
-### Low Priority
-
-- [ ] **API Reference** - Function documentation
-- [ ] **Framework Comparison** - LangGraph vs AutoGen vs CrewAI
-- [ ] **Deployment Guides** - Docker, Terraform, cloud platforms
-- [ ] **Backtesting** - Historical performance testing
-
-## 🤝 Contributing to Documentation
-
-Documentation contributions are highly valued! If you'd like to help:
-
-1. Check the [CONTRIBUTING.md](../CONTRIBUTING.md) guide
-2. Look for `*(coming soon)*` sections above
-3. Submit a PR with new documentation
-4. Ensure examples are tested and accurate
-
-**Documentation Standards:**
-
-- Use clear, concise language
-- Include code examples where applicable
-- Test all commands/scripts before documenting
-- Add screenshots/diagrams for visual learners
-- Cross-reference related documentation
-
-## 💡 Questions or Feedback?
-
-- **General Questions**: [Open a Discussion](https://github.com/rgoerwit/investment-agent-public/discussions)
-- **Documentation Issues**: [File an Issue](https://github.com/rgoerwit/investment-agent-public/issues)
-- **Missing Content**: Suggest via Issues or Discussions
-
----
-
-**Note:** This documentation is actively being developed. Check back regularly for updates, or watch the repository to get notified of new content.
