+++ src/graph/__init__.py
@@ -1,0 +1,25 @@
+"""Public graph package API."""
+
+from __future__ import annotations
+
+from .builder import create_trading_graph
+from .components import TradingContext
+from .routing import (
+    fan_out_to_analysts,
+    fundamentals_sync_router,
+    route_tools,
+    should_continue_analyst,
+    sync_check_router,
+)
+from .tool_nodes import create_agent_tool_node
+
+__all__ = [
+    "TradingContext",
+    "create_trading_graph",
+    "should_continue_analyst",
+    "route_tools",
+    "create_agent_tool_node",
+    "fan_out_to_analysts",
+    "fundamentals_sync_router",
+    "sync_check_router",
+]
